tests: added placeholder for EDITOR var
The EDITOR environmental variable may not be set when the tests are run.
However, the tests do not rely upon the editor so this should not be a problem.
In the tests the EDITOR variable is set to a placeholder value
('gist-placeholder'). This will only be local to the process running the test
and will not affect the users environment.