gist: Support running command to retrieve token
Enable running a shell command defined in the config file to retrieve
the GitHub API token.

Add the function `get_value_from_command` which handles calling the
subprocess when the `string` passed to it begins with `!`, otherwise the
original string is returned.

Updated README.rst to document this new behaviour.

Resolves #65