client: handle None filename
When creating a gist, the <filename> key will always be present in the
docopt dict. However, if it has not been set, it will take the default
value of None. If the <filename> has not been set by the user, the
default filename should be 'file1.txt'. Currently, it is set to 'null'.