client: handle user errors
The argparse.ArgumentParser class does not provide a way to intercept
errors without subclassing, so a subclass is created to raise an
exception when an error is received, e.g. if the user enters an
unsupported command or option.

I have also move the exception handling inside the 'main' function.
Perviously it was not getting called because the 'main' function is the
entry point when 'gist' is installed, which completely skipped the
try-except block.